feat: integrate ToolInvocationCard into Chat component for enhanced tool interaction and display
